LayoutTests/fast/text/attributed-substring-from-range-in-textfield-expected.txt

abcd
Testing that attributedSubstringFromRange returns a string of the correct length
The results for oversized requests are incorrect due to <rdar://problem/5400551> which causes a \n to be appended to the result.
(offset: 0  length: 0) [PASSED]: ""
(offset: 0  length: 1) [PASSED]: "a"
(offset: 0  length: 2) [PASSED]: "ab"
(offset: 0  length: 3) [PASSED]: "abc"
(offset: 0  length: 4) [PASSED]: "abcd"
(offset: 1  length: 0) [PASSED]: ""
(offset: 1  length: 1) [PASSED]: "b"
(offset: 1  length: 2) [PASSED]: "bc"
(offset: 1  length: 3) [PASSED]: "bcd"
(offset: 1  length: 4) [FAILED]: "bcd\n"
(offset: 2  length: 0) [PASSED]: ""
(offset: 2  length: 1) [PASSED]: "c"
(offset: 2  length: 2) [PASSED]: "cd"
(offset: 2  length: 3) [FAILED]: "cd\n"
(offset: 2  length: 4) [FAILED]: "cd\n"
(offset: 3  length: 0) [PASSED]: ""
(offset: 3  length: 1) [PASSED]: "d"
(offset: 3  length: 2) [FAILED]: "d\n"
(offset: 3  length: 3) [FAILED]: "d\n"
(offset: 3  length: 4) [FAILED]: "d\n"
(offset: 4  length: 0) [PASSED]: ""
(offset: 4  length: 1) [FAILED]: "\n"
(offset: 4  length: 2) [FAILED]: "\n"
(offset: 4  length: 3) [FAILED]: "\n"
(offset: 4  length: 4) [FAILED]: "\n"
